refactor: remove unused GroqApiSchemas class and clean up imports in LLMNPCSystem
